% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autoplot.R
\name{autoplot.TaskForecast}
\alias{autoplot.TaskForecast}
\title{Plot for Forecast Tasks}
\usage{
\method{autoplot}{TaskForecast}(object, ...)
}
\arguments{
\item{object}{(\link{TaskForecast}).}

\item{...}{(\code{any}):
Additional argument, passed down to the underlying \code{geom} or plot functions.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link{TaskForecast}
}
\examples{
library(mlr3)
library(mlr3forecasting)
library(ggplot2)

task = tsk("airpassengers")
autoplot(task)

task = tsk("petrol")
autoplot(task)
}
