% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureForecast.R
\name{MeasureForecast}
\alias{MeasureForecast}
\title{Forecast Regression Measure}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3:Measure]{mlr3::Measure}.}
\description{
This measure specializes \link[mlr3:Measure]{mlr3::Measure} for forecast regression problems.
The \code{task_type} is set to \code{"forecast"}.
}
\section{Construction}{
\preformatted{m = MeasureForecast$new(id, range, minimize, predict_type = "response",
     task_properties = character(0L), packages = character(0L))
}

For a description of the arguments, see \link[mlr3:Measure]{mlr3::Measure}.
The \code{task_type} is set to \code{"forecast"}.
Possible values for \code{predict_types} is "response" and "se".
}

\section{Fields}{

See \link{Measure}.
}

\section{Methods}{

See \link{Measure}.
}

\concept{Measure}
