\name{mlr3forecasting-package}
\alias{mlr3forecasting-package}
\alias{mlr3forecasting}
\docType{package}
\title{
\packageTitle{mlr3forecasting}
}
\description{
\packageDescription{mlr3forecasting}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlr3forecasting}
\packageIndices{mlr3forecasting}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mlr3forecasting}

Maintainer: \packageMaintainer{mlr3forecasting}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
