% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3temporal-package}
\alias{mlr3temporal}
\alias{mlr3temporal-package}
\title{mlr3temporal: Forecasting for 'mlr3'}
\description{
Implements tasks, learners, resamplings and measures for forcasting tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3temporal.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3temporal}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3temporal/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})

Authors:
\itemize{
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-8867-762X}{ORCID})
  \item Cornelia Gruber \email{cornelia.gruber@campus.lmu.de}
  \item Johannes Nawrath \email{johannes-nawrath@t-online.de}
  \item Ava Arshadipour \email{a.arshadipour@campus.lmu.de}
}

}
