% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingForecastHoldout.R
\name{mlr_forecast_resamplings_holdout}
\alias{mlr_forecast_resamplings_holdout}
\alias{ResamplingForecastHoldout}
\title{Holdout Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\description{
Splits data into a training set and a test set.
Parameter \code{ratio} determines the ratio of observation going into the training set (default: 2/3).
}
\section{Construction}{
\preformatted{ResamplingForecastHoldout$new()
mlr_resamplings$get("forecast.holdout")
rsmp("forecast.holdout")
}
}

\section{Fields}{

See \link{Resampling}.
}

\section{Methods}{

See \link{Resampling}.
}

\section{Parameters}{

\itemize{
\item \code{ratio} :: \code{numeric(1)}\cr
Ratio of observations to put into the training set.
}
}

\examples{
 #Create a task with 10 observations
task = mlr3::tsk("airpassengers")
task$filter(1:10)

#Instantiate Resampling
rfho = mlr3::rsmp("forecast.holdout", ratio = 0.5)
rfho$instantiate(task)

# Individual sets:
rfho$train_set(1)
rfho$test_set(1)
intersect(rfho$train_set(1), rfho$test_set(1))

# Internal storage:
rfho$instance # simple list
}
\references{
\cite{mlr3}{bischl_2012}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Resampling]{Resamplings}: \link{mlr_resamplings}

\code{as.data.table(mlr_resamplings)} for a complete table of all (also dynamically created) \link{Resampling} implementations.
}
