% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerAutoArima.R,
%   R/LearnerRegrForecastMultiv_VAR.R
\name{mlr_learners_regr.Auto.Arima}
\alias{mlr_learners_regr.Auto.Arima}
\alias{LearnerRegrForecastAutoArima}
\alias{LearnerRegrForecastMultivVAR}
\title{Auto.Arima Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\description{
A \link{LearnerRegr} for a arima model  implemented in \link[forecast:auto.arima]{forecast::auto.arima} in package \CRANpkg{forecast}.

A \link{LearnerRegrForecastMultiv} for a vector autoregressive model  implemented in \link[vars:VAR]{vars::VAR} in package \CRANpkg{var}.
}
\section{Construction}{
\preformatted{LearnerRegrForecastAutoArima$new()
mlr_learners$get("regr.rpart")
lrn("regr.rpart")
}

\preformatted{LearnerRegrForecastMultivVAR$new()
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}

\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
