% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrForecastAutoArima.R
\name{mlr_learners_regr.AutoArima}
\alias{mlr_learners_regr.AutoArima}
\alias{LearnerRegrForecastAutoArima}
\title{Auto.Arima Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\description{
A LearnerRegrForecast for an (AR)I(MA) model implemented in \link[forecast:auto.arima]{forecast::auto.arima} in package \CRANpkg{forecast}.
}
\section{Construction}{
\preformatted{LearnerRegrForecastAutoArima$new()
mlr_learners$get("regr.rpart")
lrn("regr.rpart")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
