% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureForecastMAE.R
\name{mlr_measures_forecast.mae}
\alias{mlr_measures_forecast.mae}
\alias{MeasureForecastMAE}
\title{Mean Absolute Error Measure}
\format{
\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureForecast}.
}
\description{
Mean Absolute Error Measure
}
