% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureForecastMAPE.R
\name{mlr_measures_forecast.mape}
\alias{mlr_measures_forecast.mape}
\alias{MeasureForecastMAPE}
\title{Mean Absolute Percentage Error Measure}
\format{
\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureForecastRegr}.
}
\description{
Mean Absolute Percentage Error Measure
}
