% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureForecastMSE.R
\name{mlr_measures_forecast.mse}
\alias{mlr_measures_forecast.mse}
\alias{MeasureForecastMSE}
\title{Mean Squared Error Measure}
\format{
\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureForecastRegr}.
}
\description{
Mean Squared Error Measure
}
