% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureForecast.R
\name{mlr_measures_forecast.regr}
\alias{mlr_measures_forecast.regr}
\alias{MeasureForecastRegr}
\title{General Regression Measures for Forecasting}
\format{
\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureForecast}.
}
\description{
General Regression Measures for Forecasting
}
