% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureForecastRMSE.R
\name{mlr_measures_forecast.rmse}
\alias{mlr_measures_forecast.rmse}
\alias{MeasureForecastRMSE}
\title{Root Mean Squared Error Measure}
\format{
\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureForecastRegr}.
}
\description{
Root Mean Squared Error Measure
}
