% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingForecastCV.R
\name{mlr_resamplings_forecast_cv}
\alias{mlr_resamplings_forecast_cv}
\alias{ResamplingForecastCV}
\title{Forecast Cross-Validation Resampling}
\description{
Splits data using a \code{folds}-folds (default: 10 folds) rolling window cross-validation.
}
\section{Dictionary}{

This \link{Resampling} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_resamplings} or with the associated sugar function \code{\link[=rsmp]{rsmp()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_resamplings$get("forecast_cv")
rsmp("forecast_cv")
}\if{html}{\out{</div>}}
}

\section{Parameters}{

\itemize{
\item \code{folds} (\code{integer(1)})\cr
Number of folds.
\item window_size (\code{integer(1)})\cr
(Minimal) Size of the rolling window.
\item horizon (\code{integer(1)})\cr
Forecasting horizon in the test sets.
\item fixed_window (\code{logial(1)})\cr
Flag for fixed sized window. If FALSE an expanding window is used.
}
}

\examples{
# Create a task with 10 observations
task = mlr3::tsk("airpassengers")
task$filter(1:20)

# Instantiate Resampling
rfho = mlr3::rsmp("forecast_cv", folds = 3, fixed_window = FALSE)
rfho$instantiate(task)

# Individual sets:
rfho$train_set(1)
rfho$test_set(1)
intersect(rfho$train_set(1), rfho$test_set(1))

# Internal storage:
rfho$instance #  list
}
\references{
Bergmeir, Christoph, Hyndman, J R, Koo, Bonsoo (2018).
\dQuote{A note on the validity of cross-validation for evaluating autoregressive time series prediction.}
\emph{Computational Statistics & Data Analysis}, \bold{120}, 70--83.
}
\seealso{
\itemize{
\item \code{\link[=as_benchmark_result]{as_benchmark_result()}} to convert to a \link{BenchmarkResult}.
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}:
\url{https://mlr3book.mlr-org.com/performance.html#sec-resampling}
\item Package \CRANpkg{mlr3viz} for some generic visualizations.
}

Other resample: 
\code{\link{mlr_resamplings_forecast_holdout}}
}
\concept{resample}
\section{Super class}{
\code{\link[mlr3:Resampling]{mlr3::Resampling}} -> \code{ResamplingForecastCV}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iters}}{(\code{integer(1)})\cr
Returns the number of resampling iterations, depending on the values stored in the \code{param_set}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResamplingForecastCV-new}{\code{ResamplingForecastCV$new()}}
\item \href{#method-ResamplingForecastCV-clone}{\code{ResamplingForecastCV$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="format"><a href='../../mlr3/html/Resampling.html#method-Resampling-format'><code>mlr3::Resampling$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="help"><a href='../../mlr3/html/Resampling.html#method-Resampling-help'><code>mlr3::Resampling$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="instantiate"><a href='../../mlr3/html/Resampling.html#method-Resampling-instantiate'><code>mlr3::Resampling$instantiate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="print"><a href='../../mlr3/html/Resampling.html#method-Resampling-print'><code>mlr3::Resampling$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="test_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-test_set'><code>mlr3::Resampling$test_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="train_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-train_set'><code>mlr3::Resampling$train_set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingForecastCV-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingForecastCV-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingForecastCV$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingForecastCV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingForecastCV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingForecastCV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
