% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegrForecast.R
\name{TaskRegrForecast}
\alias{TaskRegrForecast}
\title{Forecast Regression Task}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Task}/\link{TaskSupervised}/\link{TaskRegr}.}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for forecasting regression problems.
The target column is assumed to be numeric.
The \code{task_type} is set to \code{"regr"} `.
}
\section{Construction}{
\preformatted{t = TaskRegrForecast$new(id, backend, target, time.col)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier for the task.
\item \code{backend} :: \link{DataBackend}\cr
Either a \link{DataBackendLong}, a object of class \code{ts} or a \code{data.frame} with specified date column

or any object which is convertible to a DataBackend with \code{as_data_backend()}.
E.g., a object of class  \code{dts} will be converted to a \link{DataBackendLong}.
\item \code{target} :: \code{character(n)}\cr
Name of the target column(s).
\item \code{time.col} :: \code{character(1)}\cr
Name of the date column. Not needed if backend is a timeseries
}
}

\section{Fields}{

All methods from \link{TaskSupervised} and \link{TaskRegr}, and additionally:
}

\section{Methods}{

See \link{TaskSupervised} and \link{TaskRegr}.
}

\seealso{
seealso_task
}
\concept{Task}
