% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendLong.R
\name{DataBackendLong}
\alias{DataBackendLong}
\title{DataBackend}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:DataBackend]{mlr3::DataBackend}
}
\description{
DataBackend for timeseries in long-format

Data Backends provide a layer of abstraction for various data storage systems.
It is not recommended to work directly with the DataBackend.
Instead, all data access is handled transparently via the \link{Task}.

To connect to out-of-memory database management systems such as SQL servers, see \CRANpkg{mlr3db}.

The required set of fields and methods to implement a custom DataBackend is listed in the respective sections.
See \link{DataBackendDataTable} or \link{DataBackendMatrix} for exemplary implementations of the interface.
}
\section{Construction}{

Note: This object is typically constructed via a derived classes, e.g. \link{DataBackendDataTable} or \link{DataBackendMatrix},
or via the S3 method \code{\link[=as_data_backend]{as_data_backend()}}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{DataBackend$new(data, primary_key = NULL, data_formats = "data.table")
}\if{html}{\out{</div>}}
\itemize{
\item \code{data} :: \code{any}\cr
The format of the input data depends on the specialization.
E.g., \link{DataBackendDataTable} expects a \code{\link[data.table:data.table]{data.table::data.table()}} and \link{DataBackendMatrix} expects a \code{\link[Matrix:Matrix]{Matrix::Matrix()}}
constructed with the \CRANpkg{Matrix} package.
\item \code{primary_key} :: \code{character(1)}\cr
Each DataBackend needs a way to address rows, which is done via a column of unique values, referenced here by \code{primary_key}.
The use of this variable may differ between backends.
\item data_formats (\code{character()})\cr
Set of supported formats, e.g. \code{"data.table"} or \code{"Matrix"}.
}
}

\section{Fields}{

\itemize{
\item \code{nrow} :: \code{integer(1)}\cr
Number of rows (observations).
\item \code{ncol} :: \code{integer(1)}\cr
Number of columns (variables), including the primary key column.
\item \code{colnames} :: \code{character()}\cr
Returns vector of all column names, including the primary key column.
\item \code{rownames} :: (\code{integer()} | \code{character()})\cr
Returns vector of all distinct row identifiers, i.e. the primary key column.
\item \code{hash} :: \code{character(1)}\cr
Returns a unique hash for this backend. This hash is cached.
\item \code{data_formats} :: \code{character()}\cr
Vector of supported data formats.
A specific format can be chosen in the \verb{$data()} method.
}
}

\section{Methods}{

\itemize{
\item \code{data(rows = NULL, cols = NULL, format = "data.table")}\cr
(\code{integer()} | \code{character()}, \code{character()}) -> \code{any}\cr
Returns a slice of the data in the specified format.
Currently, the only supported formats are \code{"data.table"} and \code{"Matrix"}.
The rows must be addressed as vector of primary key values, columns must be referred to via column names.
Queries for rows with no matching row id and queries for columns with no matching column name are silently ignored.
Rows are guaranteed to be returned in the same order as \code{rows}, columns may be returned in an arbitrary order.
Duplicated row ids result in duplicated rows, duplicated column names lead to an exception.
\item \code{distinct(rows, cols, na_rm = TRUE)}\cr
(\code{integer()} | \code{character()}, \code{character()}, \code{logical(1)}) -> named \code{list()}\cr
Returns a named list of vectors of distinct values for each column specified.
If \code{na_rm} is \code{TRUE}, missing values are removed from the returned vectors of distinct values.
Non-existing rows and columns are silently ignored.

If \code{rows} is \code{NULL}, all possible distinct values will be returned, even if the value is not present in the data.
This affects factor-like variables with empty levels, if supported by the backend.
\item \code{head(n = 6)}\cr
\code{integer(1)} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns the first up-to \code{n} rows of the data as \code{\link[data.table:data.table]{data.table::data.table()}}.
\item \code{missings(rows, cols)}\cr
(\code{integer()} | \code{character()}, \code{character()}) -> named \code{integer()}\cr
Returns the number of missing values per column in the specified slice of data.
Non-existing rows and columns are silently ignored.
}
}

\seealso{
Extension Packages: \CRANpkg{mlr3db}
}
\concept{DataBackend}
\section{Super class}{
\code{\link[mlr3:DataBackend]{mlr3::DataBackend}} -> \code{DataBackendLong}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataBackendLong-new}{\code{DataBackendLong$new()}}
\item \href{#method-DataBackendLong-data}{\code{DataBackendLong$data()}}
\item \href{#method-DataBackendLong-head}{\code{DataBackendLong$head()}}
\item \href{#method-DataBackendLong-distinct}{\code{DataBackendLong$distinct()}}
\item \href{#method-DataBackendLong-missings}{\code{DataBackendLong$missings()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="format"><a href='../../mlr3/html/DataBackend.html#method-DataBackend-format'><code>mlr3::DataBackend$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="print"><a href='../../mlr3/html/DataBackend.html#method-DataBackend-print'><code>mlr3::DataBackend$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendLong-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendLong-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendLong$new(data, primary_key, id_col, date_col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendLong-data"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendLong-data}{}}}
\subsection{Method \code{data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendLong$data(rows, cols, data_format = "data.table", roll = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendLong-head"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendLong-head}{}}}
\subsection{Method \code{head()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendLong$head(n = 6L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendLong-distinct"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendLong-distinct}{}}}
\subsection{Method \code{distinct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendLong$distinct(rows, cols, na_rm = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendLong-missings"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendLong-missings}{}}}
\subsection{Method \code{missings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendLong$missings(rows, cols)}\if{html}{\out{</div>}}
}

}
}
