% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerForecast.R
\name{LearnerForecast}
\alias{LearnerForecast}
\title{LearnerForecast}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:Learner]{mlr3::Learner}.
}
\description{
This Learner specializes Learner for forecasting problems. The \code{task_type} is set to \code{"forecast"}.
}
\section{Construction}{
\preformatted{LearnerForecast$new(id, param_set = ParamSet$new(),
     predict_types = character(),
     feature_types = character(), properties = character(),
     packages = character())
}

For a description of the arguments, see \link[mlr3:Learner]{mlr3::Learner}.`
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
