% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureForecast.R
\name{MeasureForecast}
\alias{MeasureForecast}
\title{Forecast Regression Measure}
\description{
This measure specializes \link[mlr3:Measure]{mlr3::Measure} for forecast regression problems.
The \code{task_type} is set to \code{"forecast"}.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{m = MeasureForecast$new(id, range, minimize, predict_type = "response",
     task_properties = character(0L), packages = character(0L))
}\if{html}{\out{</div>}}

For a description of the arguments, see \link[mlr3:Measure]{mlr3::Measure}.
The \code{task_type} is set to \code{"forecast"}.
Possible values for \code{predict_types} is "response" and "se".
}

\section{Fields}{

See \link{Measure}.
}

\section{Methods}{

See \link{Measure}.
}

\concept{Measure}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureForecast}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureForecast-new}{\code{MeasureForecast$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureForecast-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureForecast-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureForecast$new(
  id,
  range,
  minimize = NA,
  average = "macro",
  aggregator = NULL,
  properties = character(),
  predict_type = "response",
  predict_sets = "test",
  task_properties = character(),
  packages = character(),
  man = NA_character_
)}\if{html}{\out{</div>}}
}

}
}
