% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3forecasting-package}
\alias{mlr3forecasting}
\alias{mlr3forecasting-package}
\title{mlr3forecasting: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
