% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrForecastAutoArima.R
\name{mlr_learners_regr.auto_arima}
\alias{mlr_learners_regr.auto_arima}
\alias{LearnerRegrForecastAutoArima}
\title{Auto.Arima Forecast Learner}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerForecast}.
}
\description{
A LearnerRegrForecast for an (AR)I(MA) model implemented in \link[forecast:auto.arima]{forecast::auto.arima} in package \CRANpkg{forecast}.
}
\section{Methods}{

See \link{LearnerForecast}, additionally:
\itemize{
\item \code{forecast(h = 10, task, new_data)}  :: \code{data.table}\cr
Returns forecasts after the last training instance.
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
