% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrForecastAverage.R
\name{mlr_learners_regr.average}
\alias{mlr_learners_regr.average}
\alias{LearnerRegrForecastAverage}
\title{Average Learner}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerForecast}.
}
\description{
A LearnerRegrForecast model based on average values implemented in \code{\link[base:mean]{base::mean()}} in package \CRANpkg{base}.
}
\section{Construction}{
\preformatted{LearnerRegrForecastAverage$new()
}
}

\section{Methods}{

See \link{LearnerForecast}, additionally:
\itemize{
\item \code{forecast(h = 10, task, new_data)}  :: \code{data.table}\cr
Returns forecasts after the last training instance.
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
