% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrForecastVAR.R
\name{mlr_learners_regr.VAR}
\alias{mlr_learners_regr.VAR}
\alias{LearnerRegrForecastVAR}
\title{Vector Autoregression Forecast Learner}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerForecast}.
}
\description{
A LearnerRegrForecast for a vector autoregressive model implemented in \link[vars:VAR]{vars::VAR} in package \CRANpkg{var}.
}
\section{Methods}{

See \link{LearnerForecast}, additionally:
\itemize{
\item \code{forecast(h = 10, task, new_data)}  :: \code{data.table}\cr
Returns forecasts after the last training instance.
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
