% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureForecastMAE.R
\name{mlr_measures_forecast.mae}
\alias{mlr_measures_forecast.mae}
\alias{MeasureForecastMAE}
\title{Mean Absolute Error Measure}
\format{
\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureForecastRegr}.
}
\description{
Mean Absolute Error Measure

Mean Absolute Error Measure
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3temporal:MeasureForecast]{mlr3temporal::MeasureForecast}} -> \code{\link[mlr3temporal:MeasureForecastRegr]{mlr3temporal::MeasureForecastRegr}} -> \code{MeasureForecastMAE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureForecastMAE-new}{\code{MeasureForecastMAE$new()}}
\item \href{#method-MeasureForecastMAE-clone}{\code{MeasureForecastMAE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureForecastMAE-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureForecastMAE-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureForecastMAE$new(id = "forecast.mae")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureForecastMAE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureForecastMAE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureForecastMAE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
