% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingForecastCV.R
\name{mlr_resamplings_forecast_cv}
\alias{mlr_resamplings_forecast_cv}
\alias{ResamplingForecastCV}
\title{Forecast Cross-Validation Resampling}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.
}
\description{
Splits data using a \code{folds}-folds (default: 10 folds) rolling window cross-validation.
}
\section{Fields}{

See \link{Resampling}.
}

\section{Methods}{

See \link{Resampling}.
}

\section{Parameters}{

\itemize{
\item \code{folds} :: \code{integer(1)}\cr
Number of folds.
\item \code{window_size} :: \code{integer(1)}\cr
(Minimal) Size of the rolling window.
\item \code{horizon} :: \code{integer(1)}\cr
Forecasting horizon in the test sets.
\item \code{fixed_window} :: \code{logial(1)}\cr
Flag for fixed sized window. If FALSE an expanding window is used.
}
}

\examples{
# Create a task with 10 observations
task = tsk("airpassengers")
task$filter(1:20)

# Instantiate Resampling
rfho = rsmp("forecast_cv", folds = 3, fixed_window = FALSE)
rfho$instantiate(task)

# Individual sets:
rfho$train_set(1)
rfho$test_set(1)
intersect(rfho$train_set(1), rfho$test_set(1))

# Internal storage:
rfho$instance #  list
}
\references{
\cite{mlr3}{bischl_2012}
paper:Ch. Bergmeir, R. J. Hyndman, B. Koo, A note on the validity of cross-validation for evaluating
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Resampling]{Resamplings}: \link{mlr_resamplings}

\code{as.data.table(mlr_resamplings)} for a complete table of all (also dynamically created) \link{Resampling} implementations.
}
