% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegrForecastAirPassengers.R
\name{mlr_tasks_airpassengers}
\alias{mlr_tasks_airpassengers}
\title{AirPassengers Forecast Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegrForecast}.}
\description{
A forecasting task for the \link[datasets:AirPassengers]{datasets::AirPassengers} data set.
}
\section{Construction}{
\preformatted{mlr_tasks$get("AirPassengers")
tsk("AirPassengers")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}

\code{as.data.table(mlr_tasks)} for a complete table of all (also dynamically created) \link[=Task]{Tasks}.
}
