% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegrForecastPetrol.R
\name{mlr_tasks_petrol}
\alias{mlr_tasks_petrol}
\title{Petrol Multivariate Forecast Class}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegrForecast}
}
\description{
A multivariate forecasting task for the \link[fma:petrol]{fma::petrol} data set.
}
\section{Construction}{
\preformatted{mlr_tasks$get("petrol")
tsk("petrol")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}

\code{as.data.table(mlr_tasks)} for a complete table of all (also dynamically created) \link[=Task]{Tasks}.
}
