% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskForecast.R
\name{TaskForecast}
\alias{TaskForecast}
\title{Forecasting Task}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Task}/\link{TaskSupervised}.
}
\description{
Introduces a new Forecasting Task that inherits from \link{TaskSupervised} which can be used to forecast time series.
This is an abstract class, which should not be instantiated.
Use the subclass \link{TaskRegrForecast} instead.
The \code{task_type} is set to \code{"forecast"}.

Note, that in case the input is a \verb{data.table' or }data.frame\verb{, }mlr3temporal\verb{expects a "wide" data.frame as input. The}tsbox::to_wide()` function can help casting time-series to this format.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{t = TaskForecast$new(id, backend, target, date_col)
}\if{html}{\out{</div>}}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier for the task.
\item \code{backend} :: \link{DataBackend}\cr
Either a \link{DataBackendLong}, an object of class \code{ts} or a \code{data.frame} with specified date column
or any object which is convertible to a DataBackendLong with \code{as_data_backend()}.
E.g., a object of class  \code{dts} (from package tsbox) will be converted to a \link{DataBackendLong}.
\item \code{target} :: \code{character(n)}\cr
Name of the target column(s).
\item \code{date_col} :: \code{character(1)}\cr
Name of the date column, only required if backend is a \code{data.frame}.
}
}

\section{Fields}{

All methods inherited from \link{TaskSupervised}, and additionally:
\itemize{
\item \code{date_col} :: \code{character(1)}\cr
Name of the date column.
}
}

\section{Methods}{

See \link{TaskSupervised}, additionally:
\itemize{
\item \code{date(row_ids = NULL)}  :: \code{data.table}\cr
(\code{integer()} | \code{character()}) -> named \code{list()}\cr
Returns the \code{date} column.
}
}

\seealso{
seealso_task

Other Task: 
\code{\link{TaskRegrForecast}},
\code{\link{mlr_tasks_airpassengers}},
\code{\link{mlr_tasks_petrol}}
}
\concept{Task}
\section{Super classes}{
\code{\link[mlr3:Task]{mlr3::Task}} -> \code{\link[mlr3:TaskSupervised]{mlr3::TaskSupervised}} -> \code{TaskForecast}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskForecast-new}{\code{TaskForecast$new()}}
\item \href{#method-TaskForecast-truth}{\code{TaskForecast$truth()}}
\item \href{#method-TaskForecast-data}{\code{TaskForecast$data()}}
\item \href{#method-TaskForecast-date}{\code{TaskForecast$date()}}
\item \href{#method-TaskForecast-clone}{\code{TaskForecast$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="add_strata"><a href='../../mlr3/html/Task.html#method-Task-add_strata'><code>mlr3::Task$add_strata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="cbind"><a href='../../mlr3/html/Task.html#method-Task-cbind'><code>mlr3::Task$cbind()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="droplevels"><a href='../../mlr3/html/Task.html#method-Task-droplevels'><code>mlr3::Task$droplevels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="filter"><a href='../../mlr3/html/Task.html#method-Task-filter'><code>mlr3::Task$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="format"><a href='../../mlr3/html/Task.html#method-Task-format'><code>mlr3::Task$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="formula"><a href='../../mlr3/html/Task.html#method-Task-formula'><code>mlr3::Task$formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="head"><a href='../../mlr3/html/Task.html#method-Task-head'><code>mlr3::Task$head()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="help"><a href='../../mlr3/html/Task.html#method-Task-help'><code>mlr3::Task$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="levels"><a href='../../mlr3/html/Task.html#method-Task-levels'><code>mlr3::Task$levels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="missings"><a href='../../mlr3/html/Task.html#method-Task-missings'><code>mlr3::Task$missings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="print"><a href='../../mlr3/html/Task.html#method-Task-print'><code>mlr3::Task$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="rbind"><a href='../../mlr3/html/Task.html#method-Task-rbind'><code>mlr3::Task$rbind()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="rename"><a href='../../mlr3/html/Task.html#method-Task-rename'><code>mlr3::Task$rename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="select"><a href='../../mlr3/html/Task.html#method-Task-select'><code>mlr3::Task$select()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="set_col_roles"><a href='../../mlr3/html/Task.html#method-Task-set_col_roles'><code>mlr3::Task$set_col_roles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="set_levels"><a href='../../mlr3/html/Task.html#method-Task-set_levels'><code>mlr3::Task$set_levels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Task" data-id="set_row_roles"><a href='../../mlr3/html/Task.html#method-Task-set_row_roles'><code>mlr3::Task$set_row_roles()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskForecast-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskForecast-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskForecast$new(id, backend, target, date_col = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskForecast-truth"></a>}}
\if{latex}{\out{\hypertarget{method-TaskForecast-truth}{}}}
\subsection{Method \code{truth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskForecast$truth(row_ids = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskForecast-data"></a>}}
\if{latex}{\out{\hypertarget{method-TaskForecast-data}{}}}
\subsection{Method \code{data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskForecast$data(rows = NULL, cols = NULL, data_format = "data.table")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskForecast-date"></a>}}
\if{latex}{\out{\hypertarget{method-TaskForecast-date}{}}}
\subsection{Method \code{date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskForecast$date(row_ids = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskForecast-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskForecast-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskForecast$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
