% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskForecast.R
\name{TaskForecast}
\alias{TaskForecast}
\title{Forecast Task}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Task}/\link{TaskSupervised}.}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for forecasting problems.
The target column is assumed to be numeric.
The \code{task_type} is set to \code{"forecast"} `.
}
\section{Construction}{
\preformatted{t = TaskRegrForecast$new(id, backend, target, date_col)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier for the task.
\item \code{backend} :: \link{DataBackend}\cr
Either a \link{DataBackendLong}, a object of class \code{ts} or a \code{data.frame} with specified date column

or any object which is convertible to a DataBackend with \code{as_data_backend()}.
E.g., a object of class  \code{dts} will be converted to a \link{DataBackendLong}.
\item \code{target} :: \code{character(n)}\cr
Name of the target column(s).
}
}

\section{Fields}{

All methods from \link{TaskSupervised}, and additionally:
\itemize{
\item \code{date_col} :: \code{character(1)}\cr
Name of the date column.
}
}

\section{Methods}{

See \link{TaskSupervised}, additionally:
\itemize{
\item \code{date(row_ids = NULL)}  :: \code{data.table}\cr
(\code{integer()} | \code{character()}) -> named \code{list()}\cr
Returns the \code{date} column.
}
}

\seealso{
seealso_task

Other Task: 
\code{\link{TaskRegrForecast}}
}
\concept{Task}
