% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskForecast.R
\name{TaskForecast}
\alias{TaskForecast}
\title{Forecasting Task}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Task}/\link{TaskSupervised}.
}
\description{
Introduces a new Forecasting Task that inherits from \link{TaskSupervised} which can be used to forecast time series.
This is an abstract class, which should not be instantiated.
Use the subclass \link{TaskRegrForecast} instead.
The \code{task_type} is set to \code{"forecast"}.

Note, that in case the input is a \verb{data.table' or }data.frame\verb{, }mlr3temporal\verb{expects a "wide"  data.frame as input. The}tsbox::to_wide()` function can help casting time-series to this format.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{t = TaskForecast$new(id, backend, target, date_col)
}\if{html}{\out{</div>}}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier for the task.
\item \code{backend} :: \link{DataBackend}\cr
Either a \link{DataBackendLong}, an object of class \code{ts} or a \code{data.frame} with specified date column
or any object which is convertible to a DataBackendLong with \code{as_data_backend()}.
E.g., a object of class  \code{dts} (from package tsbox) will be converted to a \link{DataBackendLong}.
\item \code{target} :: \code{character(n)}\cr
Name of the target column(s).
\item \code{date_col} :: \code{character(1)}\cr
Name of the date column, only required if backend is a \code{data.frame}.
}
}

\section{Fields}{

All methods inherited from \link{TaskSupervised}, and additionally:
\itemize{
\item \code{date_col} :: \code{character(1)}\cr
Name of the date column.
}
}

\section{Methods}{

See \link{TaskSupervised}, additionally:
\itemize{
\item \code{date(row_ids = NULL)}  :: \code{data.table}\cr
(\code{integer()} | \code{character()}) -> named \code{list()}\cr
Returns the \code{date} column.
}
}

\seealso{
seealso_task

Other Task: 
\code{\link{TaskRegrForecast}}
}
\concept{Task}
